requirements.txt
streamlit
streamlit-oauth
firebase-admin
python-dotenv
groq
google-auth